Add support for bunny provider
