+++ src/preferences/helpers/fontPref/index.ts
@@ -1,0 +1,3 @@
+export * from "./bunnyFonts";
+export * from "./googleFonts";
+export * from "./localFonts";
